refactor entry interface
