+++ src/CMakeLists.txt
@@ -39,8 +39,8 @@ set(SRC_FILES
         reduct/core/logger.cc
         reduct/core/error.cc
 
-        reduct/storage/async_reader.cc
-        reduct/storage/async_writer.cc
+        reduct/storage/io/async_reader.cc
+        reduct/storage/io/async_writer.cc
         reduct/storage/bucket.cc
         reduct/storage/entry.cc
         reduct/storage/storage.cc
