+++ src/reduct/storage/io/async_writer.cc
@@ -1,6 +1,6 @@
 // Copyright 2022 Alexey Timin
 
-#include "reduct/storage/async_writer.h"
+#include "async_writer.h"
 
 #include <filesystem>
 #include <fstream>
@@ -9,7 +9,7 @@
 #include "reduct/core/logger.h"
 #include "reduct/storage/block_manager.h"
 
-namespace reduct::storage {
+namespace reduct::storage::io {
 
 using core::Error;
 using google::protobuf::Timestamp;
